requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
blinker==1.9.0
blis==1.3.0
Brotli==1.1.0
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.21.1
cohere==5.14.2
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dnspython==2.7.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastavro==1.10.0
filelock==3.18.0
Flask==3.1.0
Flask-Session==0.8.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
google-auth==2.40.1
google-genai==1.15.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.27.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
murmurhash==1.0.13
mutagen==1.47.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
preshed==3.0.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==5.2.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tokenizers==0.21.1
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yt-dlp @ https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz#sha256=c5a49a855bf1e78c755b23ec081eca9795c810d93349a89782139e4b56e9ccf6
